requirements.txt
streamlit==0.72.0
torch==1.3.1
torchvision==0.4.2
Pillow==8.0.1
requests_futures
uvicorn
python-multipart